Moves VRGeneratorObject and ThermalBipole to foundation
